requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.1
cryptography==3.0
Django==3.0.8
django-rest-knox==4.1.0
djangorestframework==3.11.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
